chore: split lambda functions into a separate file
